LPS-65385 Move DestinationNames to com.liferay.portal.workflow.kaleo.runtime.constants
